Buffer now has an Out type
