+++ rust-arroyo/src/processing/stream/mod.rs
@@ -9,7 +9,7 @@ pub mod source;
 mod stage;
 
 pub use batch::stage::BatchStage;
-pub use batch::buffer::{Buffer, VecBuffer};
+pub use batch::buffer::Buffer;
 pub use batch::triggers::SizeTrigger;
 pub use collector::{NoopCollector, OffsetCollector, StreamCollector};
 pub use ext::PipelineExt;
